# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Rooming API
  version: 0.2.0
  description: Current HTTP surface exposed by the Spring controllers in this repository. Every path below is the literal controller URL path, including /api/v1 where the controller maps it.
servers:
  - url: http://localhost:8080
    description: Local Spring application root
...
security:
  - bearerAuth: []
  - accessTokenCookie: []
paths:
...
  /api/v1/broker-offices:
    get:
      tags: [Broker]
      summary: List broker offices selectable for broker additional information
      responses:
        '200':
          description: Broker office list fetched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerOfficeListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Broker]
      summary: Create a broker office selectable for broker additional information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerOfficeCreateRequest'
            example:
              officeName: SKKU Gate Realty
              officePhone: 031-299-1100
              officeAddress: 25 Seobu-ro 2126beon-gil, Jangan-gu, Suwon
      responses:
        '201':
          description: Broker office saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerOfficeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/user/broker/me/properties:
    post:
      tags: [Broker]
      summary: Create a property for a manually verified broker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerPropertyCreateRequest'
            example:
              title: Quiet one-room near SKKU main gate
              propertyType: one_room
              transactionType: MONTHLY_RENT
              depositAmount: 500
              monthlyRent: 55
              maintenanceFee: 5
              areaM2: 23.5
              floorInfo: 3F
              roomCount: 1
              bathroomCount: 1
              direction: SOUTH
              availableFrom: '2026-06-01'
              description: Window faces a side street and the entry is keypad secured.
              roadAddress: 25 Seobu-ro 2126beon-gil, Jangan-gu, Suwon
              location:
                latitude: 37.29512
                longitude: 126.97184
      responses:
        '201':
          description: Property saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerPropertyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/users/me/target-places:
    get:
      tags: [Target Places]
      summary: List the current seeker's target places
      responses:
        '200':
          description: Target place list fetched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetPlaceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Target Places]
      summary: Create a target place for the current seeker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetPlaceCreateRequest'
            example:
              category: SCHOOL
              placeName: Sungkyunkwan University Natural Sciences Campus
              roadAddress: 2066 Seobu-ro, Jangan-gu, Suwon
              location:
                latitude: 37.29445
                longitude: 126.97479
              memo: First commute destination
      responses:
        '201':
          description: Target place saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetPlaceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/users/me/target-places/{targetPlaceId}:
    put:
      tags: [Target Places]
      summary: Update one target place for the current seeker
      parameters:
        - $ref: '#/components/parameters/targetPlaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetPlaceUpdateRequest'
            example:
              category: WORK_PLACE
              placeName: SKKU Startup Support Center
              roadAddress: 2066 Seobu-ro, Jangan-gu, Suwon
              location:
                latitude: 37.29508
                longitude: 126.97311
              memo: Weekday destination
      responses:
        '200':
          description: Target place updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TargetPlaceResponse'
              example:
                success: true
                data:
                  targetPlaceId: 29
                  category: WORK_PLACE
                  placeName: SKKU Startup Support Center
                  roadAddress: 2066 Seobu-ro, Jangan-gu, Suwon
                  location:
                    latitude: 37.29508
                    longitude: 126.97311
                  memo: Weekday destination
                message: Target place updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Target Places]
      summary: Delete one target place for the current seeker
      parameters:
        - $ref: '#/components/parameters/targetPlaceId'
      responses:
        '204':
          description: Target place deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/recommendations:
    get:
      tags: [Recommendation]
      summary: List AI result snapshots saved for the current seeker
      responses:
        '200':
          description: Recommendation list fetched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Recommendation]
      summary: Get and save AI-based property recommendations
      description: Frontend sends a recommendation query here. The backend calls the AI server with ids, saves the enriched recommendation snapshots, and returns frontend-ready results.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
            example:
              query: Find a quiet one-room with a short commute to my first target place.
              preferences:
                - low maintenance fee
                - pharmacy nearby
                - subway access
              topN: 3
      responses:
        '200':
          description: Recommendations fetched and saved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/v1/recommendations/favorites:
    get:
      tags: [Recommendation]
      summary: List saved recommendation results marked favorite by the current seeker
      responses:
        '200':
          description: Favorite recommendations fetched.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteRecommendationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/recommendations/{recommendationId}:
    delete:
      tags: [Recommendation]
      summary: Delete one saved recommendation for the current seeker
      parameters:
        - $ref: '#/components/parameters/recommendationId'
      responses:
        '204':
          description: Recommendation deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/recommendations/{recommendationId}/favorite:
    post:
      tags: [Recommendation]
      summary: Mark a saved recommendation as favorite
      parameters:
        - $ref: '#/components/parameters/recommendationId'
      responses:
        '200':
          description: Recommendation marked as favorite.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Recommendation]
      summary: Remove favorite state from a saved recommendation
      parameters:
        - $ref: '#/components/parameters/recommendationId'
      responses:
        '204':
          description: Recommendation favorite removed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: ROOMING_ACCESS_TOKEN
  parameters:
...
    recommendationId:
      in: path
      name: recommendationId
      required: true
      schema:
        type: integer
        format: int64
      example: 1
    targetPlaceId:
      in: path
      name: targetPlaceId
      required: true
      schema:
        type: integer
        format: int64
      example: 10
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required.
    Forbidden:
      description: Authenticated account type is not allowed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflicts with existing data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: Upstream service failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    BrokerOfficeCreateRequest:
      type: object
      required: [officeName, officePhone, officeAddress]
      properties:
        officeName:
          type: string
          example: Rooming Realty
        officePhone:
          type: string
          example: 031-123-4567
        officeAddress:
          type: string
          example: 25 Seobu-ro 2126beon-gil, Jangan-gu, Suwon
    BrokerOfficeResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/BrokerOfficeData'
        message:
          type: string
          example: Broker office saved.
      example:
        success: true
        data:
          officeId: 4
          officeName: SKKU Gate Realty
          officePhone: 031-299-1100
          officeAddress: 25 Seobu-ro 2126beon-gil, Jangan-gu, Suwon
        message: Broker office saved.
    BrokerOfficeListResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/BrokerOfficeListData'
        message:
          type: string
          example: Broker office list fetched.
      example:
        success: true
        data:
          offices:
            - officeId: 4
              officeName: SKKU Gate Realty
              officePhone: 031-299-1100
              officeAddress: 25 Seobu-ro 2126beon-gil, Jangan-gu, Suwon
            - officeId: 9
              officeName: Hwahong Realty
              officePhone: 031-247-2200
              officeAddress: 77 Jeongjo-ro, Paldal-gu, Suwon
        message: Broker office list fetched.
    BrokerOfficeListData:
      type: object
      required: [offices]
      properties:
        offices:
          type: array
          items:
            $ref: '#/components/schemas/BrokerOfficeData'
...
    BrokerPropertyCreateRequest:
      type: object
      required: [title, depositAmount, areaM2, roadAddress, location]
      properties:
        title:
          type: string
          example: Studio near campus
        propertyType:
          type: string
          nullable: true
          example: one_room
        transactionType:
          type: string
          nullable: true
          example: MONTHLY_RENT
        depositAmount:
          type: integer
          example: 500
        monthlyRent:
          type: integer
          nullable: true
          example: 55
        maintenanceFee:
          type: integer
          nullable: true
          example: 5
        areaM2:
          type: number
          format: double
          example: 23.5
        floorInfo:
          type: string
          nullable: true
          example: 3F
        roomCount:
          type: integer
          nullable: true
          example: 1
        bathroomCount:
          type: integer
          nullable: true
          example: 1
        direction:
          type: string
          nullable: true
          example: SOUTH
        availableFrom:
          type: string
          nullable: true
          example: 2026-06-01
        description:
          type: string
          nullable: true
        roadAddress:
          type: string
          example: 123 Suwon-ro, Jangan-gu, Suwon
        location:
          $ref: '#/components/schemas/CoordinateDto'
    BrokerPropertyResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/BrokerPropertyData'
        message:
          type: string
          example: Property saved.
      example:
        success: true
        data:
          propertyId: 104
          title: Quiet one-room near SKKU main gate
          propertyType: one_room
          roadAddress: 25 Seobu-ro 2126beon-gil, Jangan-gu, Suwon
          location:
            latitude: 37.29512
            longitude: 126.97184
          depositAmount: 500
          monthlyRent: 55
          maintenanceFee: 5
          areaM2: 23.5
          floorInfo: 3F
          description: Window faces a side street and the entry is keypad secured.
          hasProperty3D: false
        message: Property saved.
...
    TargetPlaceCreateRequest:
      type: object
      description: category is the canonical request field. The controller DTO also accepts placeType as a JSON alias for category.
      required: [placeName, roadAddress, location]
      anyOf:
        - required: [category]
        - required: [placeType]
      properties:
        category:
          $ref: '#/components/schemas/PlaceCategory'
        placeType:
          description: Input alias accepted by JsonAlias("placeType").
          allOf:
            - $ref: '#/components/schemas/PlaceCategory'
        placeName:
          type: string
          example: Campus main gate
        roadAddress:
          type: string
          example: 2066 Seobu-ro, Jangan-gu, Suwon
        location:
          $ref: '#/components/schemas/CoordinateDto'
        memo:
          type: string
          nullable: true
    TargetPlaceUpdateRequest:
      type: object
      description: All fields are optional for PUT in the current DTO. placeType is accepted as an input alias for category.
      properties:
        category:
          $ref: '#/components/schemas/PlaceCategory'
        placeType:
          description: Input alias accepted by JsonAlias("placeType").
          allOf:
            - $ref: '#/components/schemas/PlaceCategory'
        placeName:
          type: string
          nullable: true
        roadAddress:
          type: string
          nullable: true
        location:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/CoordinateDto'
        memo:
          type: string
          nullable: true
    TargetPlaceListResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            targetPlaces:
              type: array
              items:
                $ref: '#/components/schemas/TargetPlaceResponseItem'
        message:
          type: string
          example: Target place list fetched.
      example:
        success: true
        data:
          targetPlaces:
            - targetPlaceId: 29
              category: SCHOOL
              placeName: Sungkyunkwan University Natural Sciences Campus
              roadAddress: 2066 Seobu-ro, Jangan-gu, Suwon
              location:
                latitude: 37.29445
                longitude: 126.97479
              memo: First commute destination
            - targetPlaceId: 31
              category: SUBWAY_STATION
              placeName: Sungkyunkwan University Station
              roadAddress: 2149 Seobu-ro, Jangan-gu, Suwon
              location:
                latitude: 37.30031
                longitude: 126.97072
              memo: Weekend transit
        message: Target place list fetched.
    TargetPlaceResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/TargetPlaceResponseItem'
        message:
          type: string
          example: Target place saved.
      example:
        success: true
        data:
          targetPlaceId: 29
          category: SCHOOL
          placeName: Sungkyunkwan University Natural Sciences Campus
          roadAddress: 2066 Seobu-ro, Jangan-gu, Suwon
          location:
            latitude: 37.29445
            longitude: 126.97479
          memo: First commute destination
        message: Target place saved.
...
    RecommendationRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          example: Find a quiet studio near campus with easy subway access.
        preferences:
          type: array
          nullable: true
          items:
            type: string
          example: [quiet, subway, low maintenance fee]
        topN:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
          default: 3
          example: 3
...
    RecommendationListResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/RecommendationListData'
        message:
          type: string
          example: Recommendation list fetched.
      example:
        success: true
        data:
          results:
            - recommendationId: 7001
              propertyId: 101
              favorite: true
              property:
                location:
                  latitude: 37.2945
                  longitude: 126.9748
                depositAmount: 500
                monthlyRent: 55
                maintenanceFee: 5
              firstTargetPlaceRoute:
                targetPlaceId: 29
                transportMode: PUBLIC_TRANSPORT
                durationMinutes: 18
                routeJson:
                  totalTime: 18
                  transferCount: 1
                  pathList: []
              infrastructures:
                - infrastructureId: 10
                  name: Campus Pharmacy
                  category: PHARMACY
                  roadAddress: 17 Seobu-ro 2126beon-gil, Jangan-gu, Suwon
                  location:
                    latitude: 37.295
                    longitude: 126.975
                  walkingMinutes: 4
              explanation: Quiet property with short commute and nearby daily services.
        message: Recommendation list fetched.
    FavoriteRecommendationResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/FavoriteRecommendationData'
        message:
          type: string
          example: Favorite recommendations fetched.
      example:
        success: true
        data:
          results:
            - recommendationId: 7001
              propertyId: 101
              favorite: true
              property:
                location:
                  latitude: 37.2945
                  longitude: 126.9748
                depositAmount: 500
                monthlyRent: 55
                maintenanceFee: 5
              firstTargetPlaceRoute: null
              infrastructures: []
              explanation: Quiet property with short commute and nearby daily services.
        message: Favorite recommendations fetched.
    RecommendationResultResponse:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/RecommendationResult'
        message:
          type: string
          example: Recommendation marked as favorite.
...